Moved handlers to a separate @ControllerAdvice annotated class
