+++ src/main/java/com/springboot/brushup/students/rest/CourseService.java
@@ -1,14 +1,10 @@
 package com.springboot.brushup.students.rest;
 
-import java.io.IOException;
 import java.util.List;
-
-import javax.servlet.http.HttpServletResponse;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -143,16 +139,4 @@ public class CourseService {
 			throw new IllegalArgumentException(msg, e);
 		}
 	}
-	
-	
-	@ExceptionHandler(NotFoundException.class)
-	void handleNotFoundException(HttpServletResponse response) throws IOException {
-	    response.sendError(HttpStatus.NOT_FOUND.value());
-	}
-	
-	@ExceptionHandler(IllegalArgumentException.class)
-	void handleIllegalArgumentException(HttpServletResponse response) throws IOException {
-	    response.sendError(HttpStatus.BAD_REQUEST.value());
-	}
-	
 }
