+++ src/main/java/com/springboot/brushup/students/rest/exceptions/RestExceptionHandler.java
@@ -1,0 +1,28 @@
+package com.springboot.brushup.students.rest.exceptions;
+
+import java.io.IOException;
+
+import javax.servlet.http.HttpServletResponse;
+
+import org.springframework.core.Ordered;
+import org.springframework.core.annotation.Order;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
+
+@Order(Ordered.HIGHEST_PRECEDENCE)
+// will cause this class to be used to handle exceptions thrown by controllers under the informed package
+@ControllerAdvice(basePackages="com.springboot.brushup.students.rest")
+public class RestExceptionHandler extends ResponseEntityExceptionHandler {
+
+	@ExceptionHandler(NotFoundException.class)
+	void handleNotFoundException(HttpServletResponse response) throws IOException {
+	    response.sendError(HttpStatus.NOT_FOUND.value());
+	}
+	
+	@ExceptionHandler(IllegalArgumentException.class)
+	void handleIllegalArgumentException(HttpServletResponse response) throws IOException {
+	    response.sendError(HttpStatus.BAD_REQUEST.value());
+	}
+}
